requirements.txt
Flask==2.0.1
Flask-Migrate==3.0.1
Flask-SQLAlchemy==2.5.1
python-dotenv==0.18.0
waitress==2.0.0
pytest==6.2.4
pyjwt==2.1.0
requests==2.26.0
flasgger>=0.9.5
psycopg2-binary
